+++ app/core/middleware/rate_limit.py
@@ -9,7 +9,7 @@ from starlette.responses import JSONResponse
 
 from app.core.config import settings
 from app.core.security import decode_access_token
-from app.db.models.rate_limit import RateLimit
+from app.db.models.auth.model import RateLimit
 
 
 def _get_identifier(request: Request, ip: str) -> str:
