+++ app/db/models/__init__.py
@@ -1,8 +1,8 @@
 """
 Export models for alembic
 """
-from .base import Base
-from .audit import Audit
-from .users import User
-from .notes import Note
-from .rate_limit import RateLimit
+from app.db.models.base import Base
+from app.db.models.audit.model import Audit
+from app.db.models.user.model import User
+from app.db.models.notes.model import Note
+from app.db.models.auth.model import RateLimit
