+++ app/services/auth/login/repository.py
@@ -3,7 +3,7 @@ Session actions
 """
 from app.core.exeptions.auth import AuthErrorHandler
 from app.core.security import generate_user_token_and_return_user
-from app.db.models.users import User
+from app.db.models.user.model import User
 from app.services.audit.repository import log_audit_event
 from app.services.logger.repository import LoggerService
