+++ app/services/user/repository.py
@@ -10,7 +10,7 @@ from sqlalchemy import or_
 
 from app.core.exeptions.user import UserErrorHandler
 from app.core.security import generate_user_token_and_return_user, decode_access_token
-from app.db.models.users import User
+from app.db.models.user.model import User
 from app.email.email_service import EmailService, EmailSchema
 from app.services.audit.repository import log_audit_event
 from app.services.logger.repository import LoggerService
