+++ test/audit.py
@@ -3,8 +3,8 @@ import bcrypt
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
 from app.db.models import Base
-from app.db.models.users import User
-from app.db.models.audit import Audit
+from app.db.models.user.model import User
+from app.db.models.audit.model import Audit
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///:memory:"
 engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=True)
